Moved initialization to Core class
