+++ Fedoraware/TeamFortress2/TeamFortress2/Core/Core.h
@@ -1,0 +1,10 @@
+#pragma once
+
+class CCore
+{
+public:
+	void Load();
+	void Unload();
+};
+
+inline CCore g_Core;
