+++ Fedoraware/TeamFortress2/TeamFortress2/DLLMain.cpp
@@ -20,134 +20,7 @@
 #include "Features/Discord/Discord.h"
 #include "Utils/Minidump/Minidump.h"
 
-void Sleep(int ms)
-{
-	std::this_thread::sleep_for(std::chrono::milliseconds(ms));
-}
-
-int StringToWString(std::wstring& ws, const std::string& s)
-{
-	const std::wstring wsTmp(s.begin(), s.end());
-	ws = wsTmp;
-
-	return 0;
-}
-
-inline void SetupDiscord()
-{
-	DiscordEventHandlers handlers = {};
-	Discord_Initialize("889495873183154226", &handlers, 0, "");
-}
-
-void InitRichPresence()
-{
-	SetupDiscord();
-	Discord_ClearPresence();
-}
-
-void ShutdownRichPresence()
-{
-	Discord_ClearPresence();
-	Discord_Shutdown();
-}
-
-void Loaded()
-{
-	I::Cvar->ConsoleColorPrintf(Vars::Menu::Colors::MenuAccent, "%s Loaded!\n", Vars::Menu::CheatName.c_str());
-	I::EngineClient->ClientCmd_Unrestricted("play vo/items/wheatley_sapper/wheatley_sapper_attached14.mp3");
-
-	const int dxLevel = g_ConVars.FindVar("mat_dxlevel")->GetInt();
-	if (dxLevel < 90)
-	{
-		MessageBoxA(nullptr, "Your DirectX version is too low!\nPlease use dxlevel 90 or higher", "dxlevel too low", MB_OK | MB_ICONWARNING);
-	}
-}
-
-void Initialize()
-{
-	g_SteamInterfaces.Init();
-	g_Interfaces.Init();
-	g_NetVars.Init();
-
-	g_Draw.RemakeFonts
-	({
-		{ 0x0, Vars::Fonts::FONT_ESP::szName.c_str(), Vars::Fonts::FONT_ESP::nTall.Value, Vars::Fonts::FONT_ESP::nWeight.Value, Vars::Fonts::FONT_ESP::nFlags.Value},
-		{ 0x0, Vars::Fonts::FONT_ESP_NAME::szName.c_str(), Vars::Fonts::FONT_ESP_NAME::nTall.Value, Vars::Fonts::FONT_ESP_NAME::nWeight.Value, Vars::Fonts::FONT_ESP_NAME::nFlags.Value },
-		{ 0x0, Vars::Fonts::FONT_ESP_COND::szName.c_str(), Vars::Fonts::FONT_ESP_COND::nTall.Value, Vars::Fonts::FONT_ESP_COND::nWeight.Value, Vars::Fonts::FONT_ESP_COND::nFlags.Value },
-		{ 0x0, Vars::Fonts::FONT_ESP_PICKUPS::szName.c_str(), Vars::Fonts::FONT_ESP_PICKUPS::nTall.Value, Vars::Fonts::FONT_ESP_PICKUPS::nWeight.Value, Vars::Fonts::FONT_ESP_PICKUPS::nFlags.Value },
-		{ 0x0, Vars::Fonts::FONT_MENU::szName.c_str(), Vars::Fonts::FONT_MENU::nTall.Value, Vars::Fonts::FONT_MENU::nWeight.Value, Vars::Fonts::FONT_MENU::nFlags.Value},
-		{ 0x0, Vars::Fonts::FONT_INDICATORS::szName.c_str(), Vars::Fonts::FONT_INDICATORS::nTall.Value, Vars::Fonts::FONT_INDICATORS::nWeight.Value, Vars::Fonts::FONT_INDICATORS::nFlags.Value},
-		{ 0x0, "Verdana", 18, 1600, FONTFLAG_ANTIALIAS},
-		{ 0x0, "Verdana", 12, 800, FONTFLAG_DROPSHADOW},
-	 });
-
-	// Initialize hooks & memory stuff
-	{
-		g_HookManager.Init();
-		g_PatchManager.Init();
-		F::NetHooks.Init();
-	}
-
-	g_ConVars.Init();
-	F::Ticks.Reset();
-
-	//F::Statistics.m_SteamID = g_SteamInterfaces.User->GetSteamID();
-
-	F::Commands.Init();
-
-	InitRichPresence();
-}
-
-void Uninitialize()
-{
-	I::EngineClient->ClientCmd_Unrestricted("play vo/items/wheatley_sapper/wheatley_sapper_hacked02.mp3");
-	G::UnloadWndProcHook = true;
-	Vars::Visuals::SkyboxChanger.Value = false;
-	Vars::Visuals::ThirdPerson.Value = false;
-
-	Sleep(100);
-
-	g_Events.Destroy();
-	g_HookManager.Release();
-	g_PatchManager.Restore();
-
-	ShutdownRichPresence();
-
-	Sleep(100);
-
-	F::Visuals.RestoreWorldModulation(); //needs to do this after hooks are released cuz UpdateWorldMod in FSN will override it
-	I::Cvar->ConsoleColorPrintf(Vars::Menu::Colors::MenuAccent, "%s Unloaded!\n", Vars::Menu::CheatName.c_str());
-}
-
-void LoadDefaultConfig()
-{
-	if (std::filesystem::exists(g_CFG.GetConfigPath() + "\\Visuals\\default.fw"))
-	{
-		g_CFG.LoadVisual("default");
-	}
-
-	Sleep(200);
-
-	if (std::filesystem::exists(g_CFG.GetConfigPath() + "\\" + g_CFG.GetCurrentConfig() + ".fw"))
-	{
-		g_CFG.LoadConfig(g_CFG.GetCurrentConfig());
-	}
-
-	Sleep(200);
-
-	g_Draw.RemakeFonts
-	({
-		{ 0x0, Vars::Fonts::FONT_ESP::szName.c_str(), Vars::Fonts::FONT_ESP::nTall.Value, Vars::Fonts::FONT_ESP::nWeight.Value, Vars::Fonts::FONT_ESP::nFlags.Value},
-		{ 0x0, Vars::Fonts::FONT_ESP_NAME::szName.c_str(), Vars::Fonts::FONT_ESP_NAME::nTall.Value, Vars::Fonts::FONT_ESP_NAME::nWeight.Value, Vars::Fonts::FONT_ESP_NAME::nFlags.Value },
-		{ 0x0, Vars::Fonts::FONT_ESP_COND::szName.c_str(), Vars::Fonts::FONT_ESP_COND::nTall.Value, Vars::Fonts::FONT_ESP_COND::nWeight.Value, Vars::Fonts::FONT_ESP_COND::nFlags.Value },
-		{ 0x0, Vars::Fonts::FONT_ESP_PICKUPS::szName.c_str(), Vars::Fonts::FONT_ESP_PICKUPS::nTall.Value, Vars::Fonts::FONT_ESP_PICKUPS::nWeight.Value, Vars::Fonts::FONT_ESP_PICKUPS::nFlags.Value },
-		{ 0x0, Vars::Fonts::FONT_MENU::szName.c_str(), Vars::Fonts::FONT_MENU::nTall.Value, Vars::Fonts::FONT_MENU::nWeight.Value, Vars::Fonts::FONT_MENU::nFlags.Value},
-		{ 0x0, Vars::Fonts::FONT_INDICATORS::szName.c_str(), Vars::Fonts::FONT_INDICATORS::nTall.Value, Vars::Fonts::FONT_INDICATORS::nWeight.Value, Vars::Fonts::FONT_INDICATORS::nFlags.Value},
-		{ 0x0, "Verdana", 18, 1600, FONTFLAG_ANTIALIAS},
-		{ 0x0, "Verdana", 12, 800, FONTFLAG_DROPSHADOW},
-	 });
-	F::Menu.ConfigLoaded = true;
-}
+#include "Core/Core.h"
 
 DWORD WINAPI MainThread(LPVOID lpParam)
 {
@@ -159,25 +32,19 @@ DWORD WINAPI MainThread(LPVOID lpParam)
 		   !GetModuleHandleW(L"stdshader_dx9.dll") ||
 		   !GetModuleHandleW(L"materialsystem.dll"))
 	{
-		Sleep(5000);
+		Sleep(2000);
 	}
 
-	Initialize();
-	LoadDefaultConfig();
-
-	g_Events.Setup({ "vote_cast", "player_changeclass", "player_connect", "player_hurt", "achievement_earned", "player_death", "vote_started", "teamplay_round_start", "player_spawn", "item_pickup" }); // all events @ https://github.com/tf2cheater2013/gameevents.txt
-
-	Loaded();
+	g_Core.Load();
 
 	while (!GetAsyncKeyState(VK_F11) || F::Menu.IsOpen)
 	{
 		Sleep(20);
 	}
 
-	Uninitialize();
+	g_Core.Unload();
 
 	FreeLibraryAndExitThread(static_cast<HMODULE>(lpParam), EXIT_SUCCESS);
-	return EXIT_SUCCESS;
 }
